댓글 작성, 조회, 수정, 삭제 프론트 구현 및 테스트

+++ src/comment/comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { Repository } from 'typeorm';
@@ -14,36 +14,56 @@ export class CommentService {
   @InjectRepository(Post)
   private postRepository: Repository<Post>;
 
-  async create(post_Id: number, createCommentDto: CreateCommentDto) {
+  async create(userId: number, post_Id: number, createCommentDto: CreateCommentDto) {
     const { content } = createCommentDto;
+    if (!content) {
+      throw new BadRequestException('내용을 입력해주세요.');
+    }
     const post = await this.postRepository.findOne({ where: { id: +post_Id } });
     if (!post) {
       throw new NotFoundException('게시글을 찾을 수 없습니다.');
     }
-    const comment = await this.commentRepository.save({ content, post });
+    const comment = await this.commentRepository.save({
+      content,
+      post,
+      user: { id: userId },
+    });
     return { message: '댓글이 생성되었습니다.', comment };
   }
 
-  async findAll() {
-    const comment = await this.commentRepository.find();
-    if (!comment) {
+  async findAll(post_Id: number) {
+    const comments = await this.commentRepository.find({
+      order: { id: 'DESC' },
+      where: { post: { id: +post_Id } },
+      relations: ['post', 'user'],
+    });
+    if (!comments) {
       throw new NotFoundException('댓글이 존재하지 않습니다.');
     }
-    return comment;
+    return comments;
   }
 
-  async findOne(id: number) {
-    const comment = await this.commentRepository.findOne({ where: { id } });
+  async findOne(post_Id: number, id: number) {
+    const comment = await this.commentRepository.findOne({
+      where: { id, post: { id: +post_Id } },
+      relations: ['post', 'user'],
+    });
     if (!comment) {
       throw new NotFoundException('댓글을 찾을 수 없습니다.');
     }
     return comment;
   }
 
-  async update(id: number, updateCommentDto: UpdateCommentDto) {
-    const comment = await this.commentRepository.findOne({ where: { id } });
+  async update(userId: number, post_Id: number, id: number, updateCommentDto: UpdateCommentDto) {
+    const comment = await this.commentRepository.findOne({
+      where: { id, post: { id: +post_Id } },
+      relations: ['user'],
+    });
     if (!comment) {
       throw new NotFoundException('댓글을 찾을 수 없습니다.');
+    }
+    if (comment.user.id !== userId) {
+      throw new UnauthorizedException('댓글을 수정할 수 있는 권한이 없습니다.');
     }
     const { content } = updateCommentDto;
     comment.content = content;
@@ -51,10 +71,16 @@ export class CommentService {
     return { message: '댓글이 수정되었습니다.', updatecomment };
   }
 
-  async remove(id: number) {
-    const comment = await this.commentRepository.findOne({ where: { id } });
+  async remove(userId: number, post_Id: number, id: number) {
+    const comment = await this.commentRepository.findOne({
+      where: { id, post: { id: +post_Id } },
+      relations: ['user'],
+    });
     if (!comment) {
       throw new NotFoundException('댓글을 찾을 수 없습니다.');
+    }
+    if (comment.user.id !== userId) {
+      throw new UnauthorizedException('댓글을 삭제할 수 있는 권한이 없습니다.');
     }
     const deletecomment = await this.commentRepository.remove(comment);
     return { message: '댓글이 삭제되었습니다.', deletecomment };
